Update patch set 8
Patch Set 8:

(11 comments)

I would prefer to split the dci roles into 2 or 3 roles. The dci-api and dci-webapp (maybe dci-db) even if it land on the same server. That will be more understable when we will want to upgrade/change a single part of the ansible script.

To sum up my wish : More modular and distributed. In ansible is just about cutting files and it help to read, patch and prepare the future.

Besides that philosophy everything is here :-) Sorry for all the comments.

Patch-set: 8
Label: Code-Review=0