Handle the split of dci cs and ui
Recently the DCI control-server and the UI have been splitted in two
different projets and are now independant.

Earlier the UI used to be attached to / directly in the control server,
this is how it was accessed. Now the UI reads /etc/dci-ui/config.json
to know the API address and makes its request there.

Change-Id: I6e9373ceb64f960fcdd7c244b03dc4583e56909f