backup: Backup the database before each upgrade
Before upgrading the dci-api package, - as this can contain model
changes - we do backup the database in case we need to rollback.

Database backup as the same version name as the dci-api package it
matches.

So for the package :

dci-api-0.0.201601190822git3382c460-1.el7.centos.x86_64

The matching SQL backup dump is

/path/to/dump/201601190822git3382c460.sql

Change-Id: I83777cb3f4b40e4b40fdd39fbd12e611ae5483a8