dci-api: Run dbsync when eventually new schema
Currently the handlers listed in

- name: Ensure DCI is installed
  yum: name={{item}} state=latest update_cache=yes
  with_items:
    - dci-common
    - dci-api
  notify:
    - reload httpd
    - run dbsync

are never called, so on upgrade of model the dbsync is never called due
to the fact that in the common role we already upgrade all the packages

- name: Ensure packages are up to their latest release
  yum: name=* state=latest update_cache=yes

Since running dbsync does not cost anything we will run it everytime
ansible provision the node. If it has to do an upgrade it will, else it
will simply do nothing.

This is a temporary solution until we bump to ansible 2.0 where we will
be able to specify an exclude array fo the `name=* state=latest`

Change-Id: If05db829480dbfbd6e518b0deea20502112f1bee