Update patch set 2
Patch Set 2: Code-Review+1

Note that the change Brad made here is not actually being tested by Jenkins, because firewall-cmd is not installed on the instances being run. My change here was to simply allow the module to be skipped if firewall-cmd is not found.

It's possible another test could be added that would test when firewalld was installed. Alternatively a new role could be created to make sure that firewalld and other deps are installed.

Patch-set: 2
Reviewer: Gerrit User 247 <247@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1