Monitoring: Change the interval of notifications
Change the interval and occurence at which the notifications are sent
out.

In its current state and with the current settings, our platform
notifies us every time a check returns there is something wrong.
More often than not, we have transient (ie. issue that raises just once
and then disapear) issues. Those issues do not need to be notified
everytime they happen as they are not 'real' issues.

As of now, the checks run every 10sec and start notifying on the first
failure. This patch proposes to run the various checks every 60sec and
be warned after the 5th failure.

Change-Id: Id42c8760bb04b74c59e917c023d80d7528dc8248