Update patch set 6
Patch Set 6:

The default path is ok for rspec. Thus we can over-right later.

My initial objective was JUST to fix the Ansible playbook after I had to do some manual changes on the production server. From my point of view, the patch is ready since the 6th and come with a rspec test.
I understand the rational behind the use of a parameter and did the change even if I'm not a big fan of the idea. Please, can we forward and get this patch merged.

Patch-set: 6