Fix postgresql configuration
Since the handler has been fixed we can now see the error related to
the fqdn/netmask issue present in the pg_hba configuration.
This patch allows to use IP address instead of fqdn to avoid useless
DNS queries. The mode and owner/group of the memory override file is
also fixed.

Change-Id: I331355691b87e8a5d147173306da00c37f143d55
Depends-On: #6354
Depends-On: #6355