Update patch set 1
Patch Set 1: Code-Review-2

I changed this because all loaded python in the WSGI is not unloaded when a new file is available. So we can be in a case with dci-api is updated and the apache server still serve old revision. It can lead to weird errors

Patch-set: 1
Reviewer: Gerrit User 107 <107@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-2