swift: Add proxy-logging middleware on proxy
Swift proxy process was consuming the same amount of ram than the
component file.
Because now OSP13 component file are ~4.5G it was using almost 60% of
the total ram for one download.

  PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND
 7463 swift     20   0 5109020   4.6g   2156 S   2.0 60.0   0:09.56 swift-proxy-server

Due to a missing proxy-logging in the middleware pipeline of the
swift proxy, the memory was consume and not released resulting in OOM

https://bugzilla.redhat.com/show_bug.cgi?id=1368641
https://review.openstack.org/#/c/340389/

Change-Id: I83d82ed793d2c5f610c74b95a430d829aebcca24