httpd: Install mod_ssl only when necessary
In non SSL environment the package mod_ssl is installed by default
causing httpd to bind on port 443. As a result in an AIO deployment
the nginx service will fail to bind on this port due to this package
installation.
This patch allow to install the mod_ssl only if the ssl_enabled flag
is set to true.

Change-Id: I3e735cc092bfd80b49a120d64090f5dc7c137d53