swift: Fix mysql module tasks
Due to the change in ansible 2.6.5 on the mysql module dependency
(switch from MySQLdb to PyMySQL) [1] we need to specify either the
user/pass or the mysql user config file.
Because we are not providing any of the previous solution and we are
using the default connection settings (via the unix socket) we need
to add this value for mysql_db and mysql_user tasks.

[1] https://github.com/ansible/ansible/commit/6e90277

Change-Id: Ia39b94fe6be0e4f63d78c4d755b9655dfb27876a