Use dci user only instead of centos
Centos user was used only for the common role which install the
dci-sshpubkeys package (with the dci user creation).
Now this package is install during the provisionning of the node so
we don't need this user anymore since there's no ssh pub key in the
centos's authorized_keys.
In the CI (dci-deploy job) the centos user was already replace by
the dci user.

Change-Id: I86bbe90e45d3f1f3139b7f7d5d2000f34afb065c