api: Remove ansible ENV variables and show output
Using environment variables in ansible could lead to sensitive
information leak when using a high verbose level.
The no_log prevents this but it also hides the output of the
command which is then hard to do debug.
Sourcing a file before le command (which contains the environment
variables) and removing the no_log fix the issue with -vvv in
ansible-playbook command.

Change-Id: I9b7e6a8b1be603c5d4246d479b4390f1d59fcd39