restart all services for dci-control-server
The dci-control-server package is used for the following services:
- dci-api (httpd)
- dci-worker

When the package is upgraded, these two services need to be
restarted/reloaded.

As these two services are handled in two different plays, if the package
is upgraded in one play, only one service is notified (notifications are
    play-scoped).

This change exports a fact and necessary triggers to be able to notify
about the package upgrade across plays.

Change-Id: I60f998ea918ce9c490a9dee963978cba8088b084