Disable auto-update for most DCI packages
Some DCI packages are already not auto-upgraded in the dci-common role
when the system-packages are updated to allow reloading or restarting
some services if these packages are upgraded.

In this change we disable generic auto-update for all DCI packges so
that we can restart services which depend on them (dci-control-server,
and later more).

- Restart httpd@dci-cs when upgrading one of:
    - dci-control-server
    - python-dciclient
    - dci-umb
    - python-dciauth

I

Change-Id: Ibbfa85b8a6b0fd7b31b19fd1d92941b9f8fc7d5b