Add max_content_length_mb var
max_content_length_mb is used to limit the size of the requests sent to
dci-api.

It was previously hardcoded as 128 in the settings.py.j2 template.

Since we have a bunch of HTTP 413 - ENTITY TOO LARGE responses we'll try
to increase this value and may have to do this a few times to find the
correct value so we introduce it as an ansible var.

Setting it intentionnally to a quite low default value (32 MB) to force
overriding it if required.

An associated changeset in dci-infra-data-prod sets it to 256 MB

Change-Id: I5187a93d424f39f8b481f058b526bfaaa7c80ba1
Depends-On: I52e2fb875ffb9505d8dfb7b7463942b74c6d8843