Disable SSL on Elastic
SSL certificate analytics.distributed-ci.io is not available for the container.
The certificate is generated automatically by the container at startup.
This raise SSL certificate problem: "unable to get local issuer certificate"
This server is behind Firewall and only consumed by the control server.
No SSL is acceptable for now.

Change-Id: I2a348f8205c7d9d7316d971788bf4b0b99ea9847