Apply the httpd role to all groups at once
Before this change, the httpd role was applied separately to each
hostgroup requiring it.

When one host was part of more than one hostgroup requiring the httpd
role, this resulted in the httpd role being applied on the same machine
as many times as the number of hostgroups.
Now the role is applied only once (saving time).

In the case of different hosts being part of these hostgroups, this
resulted in the httpd role being executed on each machine in a different
play (so, not in parallel).
Now the role is be applied to each host in parallel (saving time).

Change-Id: I36e244709d2959cd3f390b39df5acabe63ecdce6