yolo_backup.sh: Errors were not correctly handled
Errors were handled from within a sub-shell (in parentheses), the exit
statements would only exit the sub-shell, not the main script.

Running error handler in a function allows to exit from the script on
errors.

Minimal reproducer below:
```
❯ cat toto.sh; echo "==========\nEXECUTING\n====="; bash toto.sh

toto() {
  /bin/false
}

fail() {
  echo $1
  exit $2
}

toto || (
  echo "TEST 1: Error handler from a subshell (parentheses)"
  echo "will exit the subshell (withing parentheses)"
  echo "but will not exit the main script."
  exit 2
)

echo "We'd like that this would not be printed but it will."

toto || fail "TEST 2: Error handler from a function works as expected" 3

echo "This won't be printed"
==========
EXECUTING
=====
TEST 1: Error handler from a subshell (parentheses)
will exit the subshell (withing parentheses)
but will not exit the main script.
We'd like that this would not be printed but it will.
TEST 2: Error handler from a function works as expected
```

Change-Id: I850011171111f10933763cafb37319414014422c