Add role for postgresql running in a container with SSL connections
Uses the rhel9/postgresql16 container.

Uses self-signed certificates & its own CA (the CA should be made
available to all clients for verification ideally).

Connections are restricted to `apiips`.

CA is managed in a separate role + playbook, CA cert & SERVER certs are
included in some file that's ready for use in an Ansible inventory.

Change-Id: I7deb2f3e99890f42c1e08ce6097977c6ea39e18d
Depends-On: https://softwarefactory-project.io/r/c/dci-jobs/+/34422
Depends-On: https://softwarefactory-project.io/r/c/dci-control-server/+/34430